requirements.txt
pypdf
faiss-cpu
sentence-transformers
ctransformers
joblib
numpy
scipy
scikit-learn
tqdm
